fix: 에피소드 갯수를 나타내는 타입 변경

+++ src/main/java/com/yjkim/lezhin/memberContent/application/dto/TopMemberContentResult.java
@@ -3,6 +3,6 @@ package com.yjkim.lezhin.memberContent.application.dto;
 public record TopMemberContentResult(
         String title,
         String author,
-        int contentEpisodeCount
+        Long contentEpisodeCount
 ) {
 }
